# table-serve-backend/openapi.yaml
openapi: 3.1.0
info:
  title: Table Serve API
  description: |
    REST API for the Table Serve platform - an NFC-based restaurant menu and ordering system.

    Three access levels:
    - **Super Admin** - Platform administrators (Table Serve company). Requires master password to register.
    - **Admin** - Restaurant owners and managers managing their menu, tables, and analytics.
    - **Customer** - Public-facing API for NFC-scanned menu viewing and order placement.
  version: 1.0.0
  contact:
    name: Table Serve Support
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.tableserve.app
    description: Production
...
paths:
...
  /api/auth/sign-up/email:
    post:
      tags: [Auth]
      summary: Sign up with email
      operationId: signUpEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Signed up successfully
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/superadmin/register:
    post:
      tags: [Super Admin]
      summary: Register a super admin account
      description: Requires the master password configured on the server.
      operationId: registerSuperAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSuperAdminRequest'
      responses:
        '201':
          description: Super admin registered
        '403':
          description: Invalid master password
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/superadmin/stats:
    get:
      tags: [Super Admin]
      summary: Platform-wide statistics
      operationId: getSuperAdminStats
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/superadmin/organizations:
    get:
      tags: [Super Admin]
      summary: List all organizations
      operationId: listOrganizations
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of organizations
    post:
      tags: [Super Admin]
      summary: Create organization with owner
      operationId: createOrganization
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created
        '409':
          description: Slug already taken
  /api/superadmin/organizations/{id}:
    get:
      tags: [Super Admin]
      summary: Get organization details
      operationId: getOrganization
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Organization details
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Super Admin]
      summary: Delete an organization
      operationId: deleteOrganization
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Organization deleted
  /api/superadmin/organizations/{id}/suspend:
    patch:
      tags: [Super Admin]
      summary: Suspend an organization
      operationId: suspendOrganization
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Organization suspended
  /api/superadmin/organizations/{id}/activate:
    patch:
      tags: [Super Admin]
      summary: Activate a suspended organization
      operationId: activateOrganization
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Organization activated
  /api/superadmin/organizations/{id}/subscription:
    patch:
      tags: [Super Admin]
      summary: Update organization subscription
      operationId: updateSubscription
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  type: string
                  enum: [free, basic, premium]
                expiryDays:
                  type: integer
      responses:
        '200':
          description: Subscription updated
  /api/superadmin/analytics:
    get:
      tags: [Super Admin]
      summary: Platform-wide analytics
      operationId: getSuperAdminAnalytics
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          description: Analytics data
  /api/superadmin/users:
    get:
      tags: [Super Admin]
      summary: List all users
      operationId: listUsers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of users
  /api/superadmin/users/{id}/ban:
    patch:
      tags: [Super Admin]
      summary: Ban a user
      operationId: banUser
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User banned
  /api/superadmin/users/{id}/unban:
    patch:
      tags: [Super Admin]
      summary: Unban a user
      operationId: unbanUser
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User unbanned

  # ─── Admin ─────────────────────────────────
  /api/admin/register:
    post:
      tags: [Admin]
      summary: Register restaurant admin account
      operationId: registerAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAdminRequest'
      responses:
        '201':
          description: Account and organization created
        '409':
          description: Slug already taken
...
  /api/admin/analytics:
    get:
      tags: [Admin]
      summary: Restaurant analytics
      operationId: getAdminAnalytics
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          description: Analytics data
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    OrgIdHeader:
      name: x-organization-id
      in: header
      required: false
      description: Organization ID context (not required if session has active organization)
      schema:
        type: string
    PeriodQuery:
      name: period
      in: query
      schema:
        type: string
        enum: [today, week, month, quarter, year, custom]
        default: month
  responses:
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    RegisterSuperAdminRequest:
      type: object
      required: [name, email, password, masterPassword]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        masterPassword:
          type: string
    RegisterAdminRequest:
      type: object
      required: [name, email, password, organizationName, organizationSlug]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        organizationName:
          type: string
        organizationSlug:
          type: string
          pattern: '^[a-z0-9-]+$'
    CreateOrganizationRequest:
      type: object
      required: [name, slug, ownerEmail, ownerName, ownerPassword]
      properties:
        name:
          type: string
        slug:
          type: string
        ownerEmail:
          type: string
          format: email
        ownerName:
          type: string
        ownerPassword:
          type: string
          minLength: 8
...
    PlatformStats:
      type: object
      properties:
        organizations:
          type: integer
        activeOrganizations:
          type: integer
        suspendedOrganizations:
          type: integer
        users:
          type: integer
        orders:
          type: integer
        totalRevenue:
          type: string